snel-backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
